Fix behaviour with temporary queues.
It's a common behaviour in RabbitMQ to have a temporary queue created
for the purpose of Ack messaging.

When such a short lived queue is created it can be caught by the listing
of the queues, but when the stats are fetched, the queue no longer
exists.  This cause this kind of error:

    collectd[15267]: Error: HTTP Error 404: Object Not Found
    collectd[15267]: Unhandled python exception in read callback: AttributeError: 'NoneType' object has no attribute 'get'

This patch prevent this from happening in two ways:
 1. check that we have indeed some data back from rabbitmq before
 dispatching;
 2. add the ability to remove such queue from being graphed