Hack to support write_http plugin in Collectd <5.5
There's a bug with Collectd's write_http plugin that is present in
Collectd 5.4.1 (the version distributed with Ubuntu LTS). The result
is that RabbitMQ stats cannot be pushed as JSON with the write_http
plugin. This affects people publishing stats to Librato.

For more detail, see the following bug reports:

https://github.com/phobos182/collectd-elasticsearch/issues/15

https://github.com/collectd/collectd/issues/716